+++ addon/-private/message-builder.js
@@ -1,0 +1,1 @@
+export default function setMessageBuilder() {}
