+++ addon/constraints.js
@@ -1,14 +1,14 @@
-import date from './constraints/date';
-import email from './constraints/email';
-import lessThan from './constraints/less-than';
-import greaterThan from './constraints/greater-than';
-import maxLength from './constraints/max-length';
-import minLength from './constraints/min-length';
-import present from './constraints/present';
-import phoneNumber from './constraints/phone-number';
-import number from './constraints/number';
-import bigDecimal from './constraints/big-decimal';
-import truthy from './constraints/truthy';
+import date from '@zestia/ember-validation/-private/constraints/date';
+import email from '@zestia/ember-validation/-private/constraints/email';
+import lessThan from '@zestia/ember-validation/-private/constraints/less-than';
+import greaterThan from '@zestia/ember-validation/-private/constraints/greater-than';
+import maxLength from '@zestia/ember-validation/-private/constraints/max-length';
+import minLength from '@zestia/ember-validation/-private/constraints/min-length';
+import present from '@zestia/ember-validation/-private/constraints/present';
+import phoneNumber from '@zestia/ember-validation/-private/constraints/phone-number';
+import number from '@zestia/ember-validation/-private/constraints/number';
+import bigDecimal from '@zestia/ember-validation/-private/constraints/big-decimal';
+import truthy from '@zestia/ember-validation/-private/constraints/truthy';
 
 export {
   date,
