+++ tests/unit/constraints/present-test.js
@@ -1,5 +1,5 @@
 import { module, test } from 'qunit';
-import present from '@zestia/ember-validation/constraints/present';
+import { present } from '@zestia/ember-validation/constraints';
 
 module('present', function () {
   test('it returns nothing when valid', function (assert) {
